#modified:   .gitignore #modified:   Makefile #new file:   app/controllers/auth.js #new file:   app/controllers/containers.js #new file:   app/controllers/dash.js #modified:   app/controllers/entry.js #deleted:    app/controllers/error.js #deleted:    app/controllers/github.js #deleted:    app/controllers/index.js #deleted:    app/controllers/login.js #deleted:    app/controllers/logout.js #deleted:    app/controllers/profile.js #deleted:    app/controllers/signup.js #new file:   app/controllers/store.js #deleted:    app/models/index.js #modified:   app/models/user.js #new file:   app/statics/fonts/BebasNeue-webfont.eot #new file:   app/statics/fonts/BebasNeue-webfont.svg #new file:   app/statics/fonts/BebasNeue-webfont.ttf #new file:   app/statics/fonts/BebasNeue-webfont.woff #deleted:    app/statics/fonts/glyphicons-halflings-regular.eot #deleted:    app/statics/fonts/glyphicons-halflings-regular.svg #deleted:    app/statics/fonts/glyphicons-halflings-regular.ttf #deleted:    app/statics/fonts/glyphicons-halflings-regular.woff #new file:   app/statics/fonts/sourcesanspro-bold.eot #new file:   app/statics/fonts/sourcesanspro-bold.svg #new file:   app/statics/fonts/sourcesanspro-bold.ttf #new file:   app/statics/fonts/sourcesanspro-bold.woff #new file:   app/statics/fonts/sourcesanspro-light.eot #new file:   app/statics/fonts/sourcesanspro-light.svg #new file:   app/statics/fonts/sourcesanspro-light.ttf #new file:   app/statics/fonts/sourcesanspro-light.woff #new file:   app/statics/fonts/sourcesanspro-regular.eot #new file:   app/statics/fonts/sourcesanspro-regular.svg #new file:   app/statics/fonts/sourcesanspro-regular.ttf #new file:   app/statics/fonts/sourcesanspro-regular.woff #deleted:    app/statics/images/1365.2048/detroit.jpg #deleted:    app/statics/images/hapi.logo.no.name.png #deleted:    app/statics/images/hapi.logo.png #new file:   app/statics/images/hapi.logo.svg #deleted:    app/statics/javascripts/bootstrap.js #deleted:    app/statics/javascripts/bootstrap.min.js #deleted:    app/statics/javascripts/docs.min.js #new file:   app/statics/javascripts/jquery-2.1.0.js #new file:   app/statics/javascripts/jquery-2.1.0.min.js #new file:   app/statics/javascripts/modal.js #deleted:    app/statics/stylesheets/bootstrap-social.css #deleted:    app/statics/stylesheets/bootstrap-theme.css #deleted:    app/statics/stylesheets/bootstrap-theme.css.map #deleted:    app/statics/stylesheets/bootstrap-theme.min.css #deleted:    app/statics/stylesheets/bootstrap.css #deleted:    app/statics/stylesheets/bootstrap.css.map #deleted:    app/statics/stylesheets/bootstrap.min.css #deleted:    app/statics/stylesheets/cover.css #deleted:    app/statics/stylesheets/dashboard.css #deleted:    app/statics/stylesheets/docs.css #deleted:    app/statics/stylesheets/font-awesome.css #new file:   app/statics/stylesheets/fonts/bebasneue.css #new file:   app/statics/stylesheets/fonts/font-awesome.css #new file:   app/statics/stylesheets/fonts/sourcesanspro.css #new file:   app/statics/stylesheets/grid/grid.css #deleted:    app/statics/stylesheets/signin.css #new file:   app/statics/stylesheets/views/dash.css #new file:   app/statics/stylesheets/views/home.css #deleted:    app/views/404.jade #deleted:    app/views/500.jade #new file:   app/views/dash.html #new file:   app/views/entry.html #deleted:    app/views/entry.jade #deleted:    app/views/error.jade #deleted:    app/views/layout.jade #deleted:    app/views/login.jade #deleted:    app/views/profile.jade #deleted:    app/views/signup.jade #modified:   index.js #deleted:    node_modules/.gitignore #deleted:    node_modules/suit/.gitignore #deleted:    node_modules/suit/README.md #deleted:    node_modules/suit/index.js #deleted:    node_modules/suit/lib/common.js #deleted:    node_modules/suit/lib/controller.js #deleted:    node_modules/suit/lib/model.js #deleted:    node_modules/suit/lib/view.js #deleted:    node_modules/suit/package.json #deleted:    node_modules/xop/index.js #modified:   package.json
The gitignore was modified to now include the node_modules path because xop is no longer in the project directly, and has been moved to NPM officially as Slop.
The Makefile was modified to break the long command into multiple lines.
An Auth controller was added to consolidate the github, login, logout, signup, and index modules.
A Containers controller was added to allow for creation of Docker containers from an API.
A Dashboard controller was added to replace the Profile controller.
The Index controller was removed because of the dynamic requires of the controllers in the main index module.
The Store controller was added with hopes of bringing in an API store functionality.
The Error controller was removed in favor of moving thr error routes to the main index module.
The User model got new class, helper methods.
The Index model module was removed because there is only one model.
Any Bootstrap fonts were removed, and new fonts were added.
All static images were removed. There is only one image now, and it is an SVG.
Anything pertaining to Bootstrap was removed, and hand coded HTML/CSS was introduced.
jQuery was introduced as well.
A Modal Javascript file was introduced for allowing modal views in the dashboard.
All CSS files were removed, and new ones were introduced for the dashboard and homepage, and a Grid system was introduced.
All Jade files were removed. Jade has been replaced by the doT template engine. It's much harder, better, faster, stronger.
The main index module was modified to remove passport functionality into the Auth controller.
Node_modules is no longer necessary, and has been removed from the project entirely. This includes the .gitignore and Suit.
Static routes in the index module were removed in favor of dynamic route additions. Also, XOP was replaced with Slop, and the code necessary to replace Jade with doT was included.