modified:   Dockerfile
Updated the Docker version the Dockerfile was tested against.
Changed the Image name to the Paasta naming convention.
Changed the Home Directory to reflect the 'root' user since the 'git' user no longer exists.
Removed the lines that modified the default shell of the 'git' user since 'root' uses bash.
Removed the line that installs Node Forever since it's prepackaged in paasta/base.